:tractor: Refactor repo tests into their own test files
