+++ package/tests/test_repos.py
@@ -5,22 +5,6 @@ from django.test import TestCase
 from package.repos import get_repo, get_repo_for_repo_url, supported_repos
 from package.repos.base_handler import BaseHandler
 from package.repos.unsupported import UnsupportedHandler
-from package.repos.bitbucket import BitbucketHandler
-from package.repos.github import GitHubHandler
-from package.models import Package, Category, Commit
-
-
-class TestBaseHandler(TestCase):
-    def setUp(self):
-        super().setUp()
-        self.category = Category.objects.create(title="dummy", slug="dummy")
-        self.category.save()
-        self.package = Package.objects.create(
-            title="Django Piston",
-            slug="django-piston",
-            repo_url="https://bitbucket.org/jespern/django-piston",
-            category=self.category,
-        )
 
 
 def test_base_handler_not_implemented(package):
@@ -185,98 +169,7 @@ def test_get_repo_registry(package):
 # they were already commented out.
 
 
-class TestBitbucketRepo(TestBaseHandler):
-    def setUp(self):
-        super(TestBitbucketRepo, self).setUp()
-        self.package = Package.objects.create(
-            category=self.category,
-            title="django-mssql",
-            slug="django-mssql",
-            repo_url="https://bitbucket.org/Manfre/django-mssql/",
-        )
-        self.bitbucket_handler = BitbucketHandler()
-
-    def test_fetch_commits(self):
-        self.assertEqual(Commit.objects.count(), 0)
-        self.bitbucket_handler.fetch_commits(self.package)
-        self.assertNotEqual(Commit.objects.count(), 0)
-
-    def test_fetch_metadata(self):
-        package = self.bitbucket_handler.fetch_metadata(self.package)
-        self.assertTrue(
-            package.repo_description.startswith(
-                "Microsoft SQL server backend for Django running on windows"
-            )
-        )
-        self.assertTrue(package.repo_watchers > 0)
-        self.assertTrue(package.repo_forks > 0)
-        self.assertEqual(package.participants, "Manfre")
-
-
-class TestGithubRepo(TestBaseHandler):
-    def setUp(self):
-        super().setUp()
-        self.package = Package.objects.create(
-            title="Django Enhancement Proposals",
-            slug="deps",
-            repo_url="https://github.com/django/deps",
-            category=self.category,
-        )
-        self.github_handler = GitHubHandler()
-
-        self.invalid_package = Package.objects.create(
-            title="Invalid Package",
-            slug="invldpkg",
-            repo_url="https://example.com",
-            category=self.category,
-        )
-
-    def test_fetch_commits(self):
-        self.assertEqual(Commit.objects.count(), 0)
-        self.github_handler.fetch_commits(self.package)
-        self.assertTrue(Commit.objects.count() > 0)
-
-    def test_fetch_metadata(self):
-        # Currently a live tests that access github
-        package = self.github_handler.fetch_metadata(self.package)
-        self.assertEqual(
-            package.repo_description,
-            "Django Enhancement Proposals",
-        )
-        self.assertTrue(package.repo_watchers > 100)
-
-    def test_fetch_metadata_archived_repo(self):
-        # test what happens when we reference an archived repository
-        assert self.archived_package.date_repo_archived is None
-        package = self.github_handler.fetch_metadata(self.archived_package)
-
-        assert self.archived_package.date_repo_archived is not None
-        assert package.repo_description == "Django + Pagination made easy."
-        assert package.date_repo_archived is not None
-
-    def test_fetch_metadata_unsupported_repo(self):
-        # test what happens when setting up an unsupported repository
-        self.package.repo_url = "https://example.com"
-        package = self.github_handler.fetch_metadata(self.invalid_package)
-
-        self.assertEqual(package.repo_description, "")
-        self.assertEqual(package.repo_watchers, 0)
-        self.invalid_package.fetch_commits()
-        self.assertEqual(package.commit_set.count(), 0)
-
-
-class TestGitlabRepo(TestBaseHandler):
-    def setUp(self):
-        super().setUp()
-        self.package = Package.objects.create(
-            title="Django",
-            slug="django",
-            repo_url="https://gitlab.com/delta10/kees",
-            category=self.category,
-        )
-
-
-class TestRepos(TestBaseHandler):
+class TestRepos(TestCase):
     def test_repo_registry(self):
         from package.repos import get_repo, supported_repos
